DRY line item view models creation
