+++ src/Modules/OrchardCore.Commerce/Drivers/OrderPartDisplayDriver.cs
@@ -1,12 +1,13 @@
 using OrchardCore.Commerce.Abstractions;
+using OrchardCore.Commerce.Helpers;
 using OrchardCore.Commerce.Models;
-using OrchardCore.Commerce.MoneyDataType.Extensions;
 using OrchardCore.Commerce.ViewModels;
 using OrchardCore.ContentManagement;
 using OrchardCore.ContentManagement.Display.ContentDisplay;
 using OrchardCore.ContentManagement.Display.Models;
 using OrchardCore.DisplayManagement.ModelBinding;
 using OrchardCore.DisplayManagement.Views;
+using OrchardCore.Workflows.Helpers;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -65,49 +66,12 @@ public class OrderPartDisplayDriver : ContentPartDisplayDriver<OrderPart>
     private async ValueTask PopulateViewModelAsync(OrderPartViewModel model, OrderPart part)
     {
         model.ContentItem = part.ContentItem;
-        var products = await _productService.GetProductDictionaryAsync(part.LineItems.Select(line => line.ProductSku));
-        var lineItems = await Task.WhenAll(part.LineItems.Select(async lineItem =>
-        {
-            var product = products[lineItem.ProductSku];
-            var metaData = await _contentManager.GetContentItemMetadataAsync(product);
+        var lineItems = part.LineItems;
+        var lineItemViewModelsAndTotal = await OrderLineItemHelpers
+            .CreateOrderLineItemViewModelsAndTotalAsync(lineItems, _contentManager, _productService, _taxProviders);
 
-            return new OrderLineItemViewModel
-            {
-                ProductPart = product,
-                Quantity = lineItem.Quantity,
-                ProductSku = lineItem.ProductSku,
-                ProductName = product.ContentItem.DisplayText,
-                UnitPrice = lineItem.UnitPrice,
-                LinePrice = lineItem.LinePrice,
-                ProductRouteValues = metaData.DisplayRouteValues,
-                Attributes = lineItem.Attributes,
-            };
-        }));
-
-        var total = lineItems.Select(item => item.LinePrice).Sum();
-
-        if (_taxProviders.Any())
-        {
-            var taxContext = new TaxProviderContext(
-                lineItems.Select(item => new TaxProviderContextLineItem(
-                    products[item.ProductSku],
-                    item.UnitPrice,
-                    item.Quantity)),
-                new[] { total });
-
-            taxContext = await _taxProviders.UpdateWithFirstApplicableProviderAsync(taxContext);
-            total = taxContext.TotalsByCurrency.Single();
-
-            foreach (var (item, index) in taxContext.Items.Select((item, index) => (item, index)))
-            {
-                var lineItem = lineItems[index];
-                lineItem.LinePrice = item.Subtotal;
-                lineItem.UnitPrice = item.UnitPrice;
-            }
-        }
-
-        model.Total = total;
-        model.LineItems.AddRange(lineItems);
+        model.Total = lineItemViewModelsAndTotal.Total;
+        model.LineItems.AddRange(lineItemViewModelsAndTotal.ViewModels);
         model.Charges.AddRange(part.Charges);
         model.OrderPart = part;
     }
